Revert "[fluid_ops]Migrate distributed/collective/process_group.h (#65781)"
This reverts commit 45efec909f104b1be289033750147dc1df36285a.
