+++ paddle/fluid/inference/tensorrt/pir/declare_plugin.h
@@ -17,3 +17,43 @@
 #include "paddle/common/macros.h"
 
 DECLARE_FILE_SYMBOLS(generic_plugin);
+
+namespace paddle::distributed {
+
+bool ProcessGroup::Task::IsCompleted() {
+  std::lock_guard<std::mutex> lock(mutex_);
+  return is_completed_;
+}
+
+ProcessGroup::ProcessGroup(int rank, int size, int gid)
+    : rank_(rank), size_(size), gid_(gid) {
+  if (gid != kIgnoreId) {
+    auto map = ProcessGroupMapFromGid::getInstance();
+    map->insert(gid_, this);
+  }
+  const char* global_rank = std::getenv("PADDLE_TRAINER_ID");
+  PADDLE_ENFORCE_NOT_NULL(
+      global_rank,
+      phi::errors::NotFound(
+          "The environment variable 'PADDLE_TRAINER_ID' cannot be found."));
+  global_rank_ = std::atoi(global_rank);
+}
+
+// TODO(sunyilun): methods below will be removed later
+ProcessGroupIdMap& ProcessGroupIdMap::GetInstance() {
+  static ProcessGroupIdMap instance;
+  return instance;
+}
+
+void ProcessGroupIdMap::DestroyProcessGroup() {
+  auto& id_map = ProcessGroupIdMap::GetInstance();
+  for (auto& item : id_map) {
+    auto use_count = item.second.use_count();
+    for (int i = 0; i < use_count; ++i) {
+      item.second.reset();
+    }
+  }
+  id_map.clear();
+}
+
+}  // namespace paddle::distributed
